requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0

# Message Queue
redis==5.0.1
celery==5.3.4

# Database
pymongo==4.6.1
motor==3.3.2

# Translation
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
sacremoses==0.1.1

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Validation
pydantic==2.5.3
python-dotenv==1.0.0

# Utilities
pybreaker==1.0.1
python-json-logger==2.0.7
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1

